sms2mqtt-persistence/requirements.txt
# SMS2MQTT Persistence — MQTT listener to PostgreSQL
paho-mqtt>=2.0,<3
psycopg2-binary>=2.9,<3
certifi>=2023.7.22
pytest>=7.0
